Refactor to reuse existing components instead of redeveloping functionality
